refactor: Split integration tests
Split tests into multiple files, since singe lib.rs became too big.
